Moves  to test common package.
