--- go/common/log/logutil/README.md
@@ -1,1 +1,0 @@
-This package contains utils related to logging.
